requirements.txt
aiogram==3.13.1
telethon==1.37.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.9.2